fix administrator administrator forget pwd

+++ jcart-store-back/src/main/java/io/cjc/jcartstoreback/dto/in/ReturnApplyInDTO.java
@@ -13,7 +13,7 @@ public class ReturnApplyInDTO {
     private Boolean opened;
     private String comment;
 
-    public Long getOrderId() {
+    public Integer getOrderId() {
         return orderId;
     }
